+++ controller/manageStore/deleteStore.js
@@ -1,21 +1,34 @@
 // const { store } = require('../../models');
 module.exports = {
   get: async (req, res) => {
-    const { store } = require('../../models');
+    const { user, store } = require('../../models');
+    const jwt = require('jsonwebtoken');
+    const access_token = req.headers['x-access-token'];
+
     if (
       req.body === undefined ||
-      req.body.userId === undefined ||
       req.body.storeId === undefined
     ) {
       res.status(404).send('Bad Request');
     }
-    const { userId, storeId } = req.body;
+    const { storeId } = req.body;
     console.log('바디바디', req.body);
+
+    const decoded = jwt.decode(access_token, { complete: true });
+    const account = decoded.payload.account;
+    const id = await user.findOne({
+      where : { email : account},
+      attributes: ['id']
+    })
+    const checkStore = await store.findOne({where : { id : storeId, userId : id.dataValues.id}})
+    console.log(id.dataValues.id, checkStore)
+    if(!checkStore) return res.status(404).json('해당 매장은 접근 권한이 없습니다')
+
     await store
       .destroy({
         where: {
           id: storeId,
-          userId: userId,
+          userId: id.dataValues.id,
         },
       })
       .then((result) => {
